+++ Source/SmithUE/Public/Commands/SmithUEUMGCommands.h
@@ -14,4 +14,5 @@ private:
 	static TSharedPtr<FJsonObject> HandleReadWidgetBlueprint(const TSharedPtr<FJsonObject>& Params);
 	static TSharedPtr<FJsonObject> HandleAddWidget(const TSharedPtr<FJsonObject>& Params);
 	static TSharedPtr<FJsonObject> HandleSetWidgetProperty(const TSharedPtr<FJsonObject>& Params);
+	static TSharedPtr<FJsonObject> HandleBindWidgetEvent(const TSharedPtr<FJsonObject>& Params);
 };